L-Variant/producers/modbus/requirements.txt
kafka-python~=2.0.2
pyModbus==2.5.3
faust